Adding clang-tidy to the cmake file, disabled for now until issues are resolved. Fixed clang-tidy warnings in system code using reserved names, since we're implementing said reserved names.
