+++ kernel_stdlib/src/typeinfo.cpp
